fix(ui-library): rename tooltip and create example folder (#860)
