+++ packages/ui-library/src/components/tooltip/index.stories.ts
@@ -1,46 +1,10 @@
-// this loads the all components instances and registers their html tags
+import { html } from 'lit';
+
 import '../../index';
-import { html } from 'lit-html';
-import { BlrTooltipType } from './index';
+import { BlrTooltipType } from '.';
+
 import { BlrTooltipRenderFunction } from './renderFunction';
 import { Themes } from '../../foundation/_tokens-generated/index.themes';
-import { TooltipPlacement } from '../../globals/constants';
-
-const sharedStyles = html`
-  <style>
-    .wrapper {
-      margin: 1.25rem;
-      display: flex;
-      flex-direction: column;
-      flex-basis: 25%;
-    }
-    .container {
-      height: 100px;
-      margin: 30px auto 0;
-      width: 200px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-    .tooltip-container {
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: center;
-    }
-    .blue-box {
-      height: 50px;
-      width: 100px;
-      background-color: lightblue;
-    }
-    .label {
-      font-family: Source Sans Pro, sans-serif;
-      font-weight: 400;
-      line-height: 1rem;
-      font-size: 1rem;
-      text-align: center;
-    }
-  </style>
-`;
 
 export default {
   title: 'Components/Tooltip',
@@ -52,14 +16,7 @@ export default {
         category: 'Appearance',
       },
     },
-    placement: {
-      description: 'Select placement of the component to the related element.',
-      options: TooltipPlacement,
-      control: { type: 'select' },
-      table: {
-        category: 'Appearance',
-      },
-    },
+
     elevation: {
       description: 'Choose if the component should be elevated.',
       control: { type: 'boolean' },
@@ -67,15 +24,15 @@ export default {
         category: 'Appearance',
       },
     },
-    offset: {
-      description: 'Enter the offset between the tooltip and the related element.',
-      control: { type: 'number' },
+    hasArrow: {
+      description: 'Choose if component has an arrow, pointing to the related element.',
+      control: { type: 'boolean' },
       table: {
         category: 'Appearance',
       },
     },
-    hasArrow: {
-      description: 'Choose if component has an arrow, pointing to the related element.',
+    visible: {
+      description: 'Toggles the visibility of the component.',
       control: { type: 'boolean' },
       table: {
         category: 'Appearance',
@@ -88,6 +45,14 @@ export default {
         category: 'Content / Settings',
       },
     },
+    static: {
+      description:
+        'Select if the component position should be static or dynamic when used inside the Tooltip component with Floating UI.',
+      control: { type: 'boolean' },
+      table: {
+        category: 'Technical Attributes',
+      },
+    },
   },
 
   parameters: {
@@ -97,22 +62,17 @@ export default {
       url: 'https://www.figma.com/file/C4vgEKz8mKyulJ4gm3Qdql/%F0%9F%AB%A7-%5BBLR%5D-The-B01LER?node-id=3618%3A126744&mode=dev',
     },
     viewMode: 'docs',
-    layout: 'padded',
+    layout: 'centered',
     docs: {
       description: {
         component: `
   <markdown>
-  Tooltip is triggered by hovering the mouse cursor over the element. It is designed to be unobtrusive and usually consist
-  of a short snippet of text. Tooltip component utilizes Floating UI library for granular positioning and interactions
-  based on user events.
+  Tooltip is triggered by hovering the mouse cursor over the element. It is designed to be unobtrusive and usually consist of a short snippet of text. To make this work technically the tooltip component can be combined with [Floating UI](https://floating-ui.com/) like shown on the example page [Tooltip using Floating UI](?path=/docs/examples-tooltip-using-floating-ui--docs).
   
   - [**Appearance**](#appearance)
-    - [**Placement**](#placement)
     - [**Elevation**](#elevation)  
     - [**Has Arrow**](#has-arrow)
-    - [**Offset**](#offset)
-  - [**Dependencies**](#dependencies)
-    - [**Tooltip Bubble**](#tooltip-bubble)
+
   </markdown>
 `,
       },
@@ -120,61 +80,29 @@ export default {
   },
 };
 
-export const Tooltip = (params: BlrTooltipType) => html` <div class="container">
-  ${BlrTooltipRenderFunction(params, html`<div class="blue-box"></div>`)}
+export const Tooltip = (params: BlrTooltipType) => html`<div style="position: relative; height: 100px;">
+  ${BlrTooltipRenderFunction(params)}
 </div>`;
 
 const defaultParams: BlrTooltipType = {
   theme: 'Light',
-  placement: 'top',
   elevation: true,
   hasArrow: true,
-  offset: 4,
+  visible: true,
   message: 'Message-text',
+  static: true,
 };
 
 Tooltip.args = defaultParams;
 
 /**
  * ## Appearance
- * ### Placement
- * The Tooltip component can be positioned in any of the following placements: top, top-start, top-end, right, right-start, right-end, bottom, bottom-start, bottom-end, left, left-start, and left-end.
- */
-
-export const Placement = () => {
-  const tooltips: BlrTooltipType[] = [
-    { ...defaultParams, message: 'Top' },
-    { ...defaultParams, placement: 'top-start', message: 'Top-start' },
-    { ...defaultParams, placement: 'top-end', message: 'Top-end' },
-    { ...defaultParams, placement: 'right', message: 'Right' },
-    { ...defaultParams, placement: 'right-start', message: 'Right-start' },
-    { ...defaultParams, placement: 'right-end', message: 'Right-end' },
-    { ...defaultParams, placement: 'bottom', message: 'Bottom' },
-    { ...defaultParams, placement: 'bottom-start', message: 'Bottom-start' },
-    { ...defaultParams, placement: 'bottom-end', message: 'Bottom-end' },
-    { ...defaultParams, placement: 'left', message: 'Left' },
-    { ...defaultParams, placement: 'left-start', message: 'Left-start' },
-    { ...defaultParams, placement: 'left-end', message: 'Left-end' },
-  ];
-
-  return html`
-    ${sharedStyles}
-
-    <div class="tooltip-container">
-      ${tooltips.map((tooltip) => html` <div class="wrapper">${Tooltip(tooltip)}</div> `)}
-    </div>
-  `;
-};
-
-Placement.story = { name: ' ' };
-
-/**
+* ### Elevation
 With elevation, the Tooltip component can be given a lifted appearance with subtle shadows or kept flat for a more minimalist design. 
  */
 
 export const Elevation = () => {
   return html`
-    ${sharedStyles}
     ${Tooltip({
       ...defaultParams,
       message: 'With elevation',
@@ -186,14 +114,14 @@ export const Elevation = () => {
     })}
   `;
 };
+Elevation.story = { name: ' ' };
 
 /**
- * The Tooltip component can have a pointing arrow, directing the user's attention to the specific element, or appear as a self-contained box.
+ * The Tooltip component can have a pointing arrow, directing the user's attention to the specific element, or appear as a self-contained box. [Floating UI](https://floating-ui.com/) can also be utilized to change the direction of the arrow and always point to the referenced element like shown on the example page [Tooltip using Floating UI](?path=/docs/examples-tooltip-using-floating-ui--docs).
  */
 
 export const HasArrow = () => {
   return html`
-    ${sharedStyles}
     ${Tooltip({
       ...defaultParams,
       message: 'With arrow',
@@ -205,46 +133,3 @@ export const HasArrow = () => {
     })}
   `;
 };
-
-/**
- * The positioning of the Tooltip component can be adjusted with an offset from the associated element.
- */
-
-export const Offset = () => {
-  return html`
-    ${sharedStyles}
- 
-      ${Tooltip({
-        ...defaultParams,
-        offset: 0,
-        placement: 'right',
-        message: 'Without offset',
-      })}
-    </div>
-
-    ${Tooltip({
-      ...defaultParams,
-      offset: 50,
-      placement: 'right',
-      message: 'With offset',
-    })}
-    
-  `;
-};
-/**
- * ## Dependencies
- * ### Tooltip Bubble
- * The Tooltip component makes use of the Tooltip Bubble component. For more information have a look at the [Tooltip Bubble](?path=/docs/design-system-web-components-feedback-tooltip-tooltip-bubble--docs) component.
- */
-
-export const TooltipBubble = () => {
-  return html`
-    ${sharedStyles}
-    ${Tooltip({
-      ...defaultParams,
-      message: 'This is the tooltip bubble',
-    })}
-  `;
-};
-
-TooltipBubble.story = { name: ' ' };
