+++ packages/ui-library/src/components/tooltip/index.test.ts
@@ -4,30 +4,19 @@ import { BlrTooltipRenderFunction } from './renderFunction';
 import type { BlrTooltipType } from '.';
 
 import { fixture, expect } from '@open-wc/testing';
-import { html } from 'lit';
 import { querySelectorDeep } from 'query-selector-shadow-dom';
 
 const sampleParams: BlrTooltipType = {
   message: 'Tooltip text comes here Tooltip text comes here',
-  placement: 'right',
+  static: true,
 };
 
-const testContent = html`<div
-  className="blue-box"
-  style="height: 200px; width: 200px; background-color: lightblue"
-></div>`;
-
 describe('blr-tooltip', () => {
-  it('is having a tooltip bubble element', async () => {
-    const element = await fixture(BlrTooltipRenderFunction(sampleParams, testContent));
+  it('is having a tooltip element', async () => {
+    const element = await fixture(BlrTooltipRenderFunction(sampleParams));
 
-    const tooltip = querySelectorDeep('blr-tooltip-bubble', element.getRootNode() as HTMLElement);
+    const tooltip = querySelectorDeep('blr-tooltip', element.getRootNode() as HTMLElement);
 
     expect(tooltip).to.exist;
-  });
-
-  it('is rendering the tooltip child element', async () => {
-    const element = await fixture(BlrTooltipRenderFunction(sampleParams, testContent));
-    expect(element.childNodes[1]).to.exist;
   });
 });
