+++ packages/ui-library/src/components/tooltip/indexReact.ts
@@ -3,7 +3,7 @@ import { createComponent } from '@lit-labs/react';
 
 import { BlrTooltip } from '.';
 
-export const BlrToolTipReact = createComponent({
+export const BlrTooltipReact = createComponent({
   tagName: 'blr-tooltip',
   elementClass: BlrTooltip,
   react: React,
