+++ packages/ui-library/src/examples/tooltip-using-floating-ui/index.css.ts
@@ -1,0 +1,7 @@
+import { typeSafeNestedCss } from "../../utils/nested-typesafe-css-literals";
+
+export const styleCustom = typeSafeNestedCss`
+  :host {
+    display: inline-block;
+  }
+`;
