+++ packages/ui-library/src/examples/tooltip-using-floating-ui/indexReact.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+import { createComponent } from '@lit-labs/react';
+
+import { BlrTooltipUsingFloatingUI } from '.';
+
+export const BlrToolTipUsingFloatingUIReact = createComponent({
+  tagName: 'blr-tooltip-using-floating-ui',
+  elementClass: BlrTooltipUsingFloatingUI,
+  react: React,
+});
